RAD-255 fix pixmap_stepsize type from float to integer (#2201)

+++ romancal/resample/resample_step.py
@@ -70,7 +70,7 @@ class ResampleStep(RomanStep):
         good_bits = string(default='~DO_NOT_USE+NON_SCIENCE')  # The good bits to use for building the resampling mask.
         include_var_flat = boolean(default=False)  # include var_flat in output image
         propagate_dq = boolean(default=False)  # propagate DQ during resampling
-        pixmap_stepsize = float(default=10)  # step size for computation of the pixel map
+        pixmap_stepsize = integer(default=10)  # step size for computation of the pixel map
         pixmap_order = integer(1, 3, default=3)  # interpolating spline order (1 or 3) used when pixmap_stepsize > 1
     """
